Convert results to int

+++ data/predictions-views.sql
@@ -2,11 +2,11 @@ DROP MATERIALIZED VIEW IF EXISTS joint_rows_predictions_geom CASCADE;
 create MATERIALIZED view joint_rows_predictions_geom
 AS
 select te.kod,
-       te.trenovacitypkod train_kod,
-       coalesce(te.finaltypkod, te.predikce) final_kod,
+       te.trenovacitypkod::int train_kod,
+       coalesce(te.finaltypkod, te.predikce)::int final_kod,
        ru.geom,
-       te.finaltypkod alg_kod,
-       te.predikce ml_kod
+       te.finaltypkod::int alg_kod,
+       te.predikce::int ml_kod
 from joint_rows_predictions te
 inner join joint_rows_ruian ru on ru.kod = te.kod
 ;
